+++ src/voidgl/base-layers/circles/circle-layer.ts
@@ -2,6 +2,7 @@ import * as Three from "three";
 import { InstanceProvider } from "../../instance-provider";
 import { Bounds, IPoint } from "../../primitives";
 import { ILayerProps, IModelType, Layer } from "../../surface/layer";
+import { LayerBufferType } from "../../surface/layer-processing/layer-buffer-type";
 import {
   IMaterialOptions,
   InstanceAttributeSize,
@@ -165,7 +166,13 @@ export class CircleLayer<
           name: "layerOpacity",
           size: UniformSize.ONE,
           update: (_uniform: IUniform) => [
-            this.props.opacity === undefined ? 1.0 : this.props.opacity
+            (this.props.opacity === undefined ? 1.0 : this.props.opacity) *
+              // HACK: There is a blending issue with three OR with WebGL itself. This is an adjustment
+              // to address this issue.
+              this.bufferType ===
+            LayerBufferType.INSTANCE_ATTRIBUTE_PACKING
+              ? 0.15
+              : 1.0
           ]
         }
       ],