+++ test/backend/test-modules/no-fragment-0.fs
@@ -1,0 +1,3 @@
+${import: no-fragment-1}
+
+// No-fragment-0 FS